# made env._compute_observations() faster