replaced the expensive tile window query (on row and col) with the index look up. 

 - tile map window query: 0.1153785250025976 s per 1000 calls
 - look up by index: 0.004833787003008183 s per 1000 calls
 - A test was added to check these outputs match: `tests/core/test_observation_tile.py`

so, `env._compute_observations()` got much faster, from 2.03 s -> 0.65 s per 100 calls

before: (WSL, the default task)
 - env.step({}): 7.819248097999662
 - env.realm.step(): 1.601971907002735
 - env._compute_observations(): 2.038284816997475
 - obs.to_gym(), ActionTarget: 1.49435229300434
 - env._compute_rewards(): 1.6132002679951256

after:
 - env.step({}): 6.491634746002092
 - env.realm.step(): 1.611555720002798
 **- env._compute_observations(): 0.647985998999502**
 - obs.to_gym(), ActionTarget: 1.5792435760013177
 - env._compute_rewards(): 1.5641796529998828